put all form controller into separate files and fixed modals to work for adding or editing tests
