+++ app/js/tests/controllers/test_controller.js
@@ -78,107 +78,10 @@ module.exports = function(app) {
     };
 
     $scope.newTest = function() {
-      var controller = ['$scope', '$modalInstance', 'dataStore', function($scope, $modalInstance, dataStore) {
-
-
-        $scope.buttonText = 'Add Test';
-        $scope.question;
-        $scope.areWeAddingQuestions;
-        $scope.standard = dataStore.standard;
-        $scope.$on('test:changed', function(event, data) {
-          $scope.test = data.test;
-        });
-
-        $scope.setData = function(test) {
-          if (!test) {
-            if (!dataStore.test) {
-              test = {
-                questions: []
-              };
-              dataStore.test = test;
-            } else {
-              if(!dataStore.test.questions) {
-                dataStore.test.quesions = [];
-              }
-              test = dataStore.test;
-            }
-          }
-          setGoal(test);
-        };
-
-        var setGoal = function(test) {
-          test._goal = test.goalId;
-        };
-
-
-        $scope.cancel = function(test) {
-          angular.copy(dataStore.masterTest, test);
-          $scope.setGoal(test);
-          if ($scope.isTestFormShowing) {
-            $scope.toggleTestForm();
-          }
-        };
-
-        $scope.addNewTest = function(test) {
-          test.standardId = $scope.standard._id;
-          $scope.createTest(test);
-        };
-
-        $scope.createTest = function(test) {
-          var newTest = angular.copy(test);
-          var numberOfTests = $filter('filter')(dataStore.tests, {standardId: $scope.standard._id});
-          test = {};
-          newTest.testName = 'Test ' + (numberOfTests.length + 1);
-          dataStore.tests.push(newTest);
-          dataStore.createTest(newTest, function(err, data) {
-            if (err) {
-              return dataStore.errors.push({
-                'msg': 'There was an error creating your test'
-              });
-            }
-
-            $scope.test = null;
-            dataStore.tests.splice(dataStore.tests.indexOf(newTest), 1, data);
-            $scope.tests = dataStore.tests;
-
-          });
-          $modalInstance.close();
-        };
-
-        $scope.newQuestions = function(test) {
-          $scope.areWeAddingQuestions = true;
-          if (!test.questions) {
-            test.questions = [];
-          }
-        };
-
-        $scope.cancelQuestion = function(question) {
-          // angular.copy($scope.master, question);
-          // question.editing = false;
-          // $scope.areWeAddingQuestions = false;
-          console.log(question);
-        };
-
-        $scope.cancel = function(test) {
-          dataStore.test = null;
-          $modalInstance.dismiss();
-        };
-
-        $scope.addQuestion = function(question) {
-          question.answers = Object.keys(question.answers)
-            .map(function(key) {
-              return question.answers[key];
-            });
-          dataStore.test = $scope.test;
-          dataStore.addQuestion(question);
-          $scope.question = {};
-          $scope.areWeAddingQuestions = false;
-        };
-      }];
       $modal.open({
         animation: $scope.animationsEnabled,
         templateUrl: '/templates/directives/test_form.html',
-        controller: controller,
+        controller: 'AddTestCtrl',
         size: 'lg'
       });
     };
@@ -201,13 +104,17 @@ module.exports = function(app) {
     };
 
     $scope.editTest = function(test) {
-      test.editing = true;
-      dataStore.test = test;
-      dataStore.masterTest = angular.copy(test);
-      $scope.test = dataStore.test;
-      if (!$scope.isTestFormShowing){
-        $scope.toggleTestForm();
-      }
+      $modal.open({
+        animation: $scope.animationsEnabled,
+        templateUrl: '/templates/directives/test_form.html',
+        controller: 'EditTestCtrl',
+        size: 'lg',
+        resolve: {
+          test: function() {
+            return test;
+          }
+        }
+      });
     };
 
     $scope.toggleDelete = function() {
