+++ app/js/tests/directives/pc_single_test.js
@@ -20,7 +20,8 @@ module.exports = function(app) {
         toggleSingleTest: '&',
         toggleTestForm: '&',
         isTestShowing: '=',
-        isTestFormShowing: '='
+        isTestFormShowing: '=',
+        editTest: '&'
       },
       controller: 'testCtrl'
     };
