+++ app/templates/directives/test_form.html
@@ -1,5 +1,5 @@
 <form class=""
-  data-ng-submit="addNewTest(test)"
+  data-ng-submit="save(test)"
   data-ng-init="setData(test)">
   <header class="modal-header">
     <h3>{{test.testName || 'New Test'}} </h3>
